更新Encoder部分代码,添加了一些代码注释

+++ JPEGCodec/Quantizer.cpp
@@ -1,6 +1,6 @@
 #include "Quantizer.h"
 
-DWORD Quantizer::_std_qtable_Y[8][8] = {
+BYTE Quantizer::_std_qtable_Y[8][8] = {
 		16,11,10,16,24,40,51,61,
 		12,12,14,19,26,58,60,55,
 		14,13,16,24,40,57,69,56,
@@ -11,7 +11,7 @@ DWORD Quantizer::_std_qtable_Y[8][8] = {
 		72,92,95,98,112,100,103,99
 };
 
-DWORD Quantizer::_std_qtable_C[8][8] = {
+BYTE Quantizer::_std_qtable_C[8][8] = {
 	17,18,24,47,99,99,99,99,
 	18,21,26,66,99,99,99,99,
 	24,26,56,99,99,99,99,99,
